Removed call to setup, which can cause problems
This script was calling setup(), which all other scripts do - but since it's a script to register and only uses the svc account credentials, it doesn't make sense to print who you are authed is.  Someone was running this with an account that wasn't registered with FC and got an error, but it would have run if setup was not called.